mesa: use ccache to speed up Mesa building
Split up Mesa image building in two steps: one for preparing it to
compile (docker build) and another one to build mesa (docker run).

This is required because we want to share an external .ccache directory,
but this can't be done in the build, but in the run.

Later we will use docker commit to tag the final image.

This fixes #issue 5.