Use a packaged LLVM 3.3 version
Unfortunately, LLVM 3.3 is not packaged for Ubuntu Xenial. And
porting from old Ubuntu Trusty is causing lot of troubles.

So we were just building LLVM 3.3 from scratch. But this takes time.
Fortunately, we can package that manually-built version directly using
checkinstall.

This commit just uses LLVM 3.3 from a package created with checkinstall.
It contains the same version as if it were build manually.

This fixes issue #15.