Port to Rocker
Rocker is an utility to create docker images. It is like a docker build
with steroids.

Rocker uses rockerfiles, which are compatible with dockerfiles, but
provides more options.

Among the options it provides, the main two that justifies the change
are:

- Support for templates, using Go's text/template. So no more need to
  use Jinja
- Allow to mount directories on build. So we can mount and share
  ~/.ccache, without needing to split the build in several steps.

It provides some other nice features that we use, but the above ones
are the most important ones.