mesa: enable proper drivers depending on LLVM version
The drivers to be built depend on the used LLVM version. But this can
change from one Mesa version to another.

Instead of hardcoding the list of drivers to build, let's parse
`configure.ac` and compare current LLVM version with the minimum
required for each driver, enabling or disabling them accordingly.

This affects to Vulkan and Gallium drivers.