llvm: define LD_LIBRARY_PATH for LLVM 3.3
For Ubuntu packages, LLVM already put (part of the) .so files in the
system library path, and then create soft links.

But for our LLVM 3.3, everything is put in the same directory. So
afterwards, it doesn't find libLLVM-3.3.so when running Mesa llvmpipe
tests.

To make things easier, just set up LD_LIBRARY_PATH to point to the place
where all the .so are put.