base: explicitly add LDFLAGS
As some of the packages are not installed in standard place (they are in
/usr/local/lib), there are situations where build is failing due not
finding the proper libraries.

This is specially the case for `make distcheck`, which fails in our
system due this situation.

A workaround that seems to work fine is to export LDFLAGS to
`-L/usr/local/lib`.